+++ CHANGELOG.md
@@ -12,6 +12,7 @@ All notable changes to this project will be documented in this file.
 - Description: Laikago MJCF
 - Description: Panda MJCF
 - Description: UR5e MJCF
+- Loader: Pinocchio
 
 ### Changed